Factor out log wrangling to a module
